# Refactor multiply operation for better performance in 2 Commands.Utility functions