@iSazonov 2020-11-20
@xtqqczze Not for the PR: could you please fix `!(this is` pattern in ProviderBase.cs.