@iSazonov 2020-11-20
> there is in fact a very small gain.

0.0358 ns vs 0.6288 ns small?