@xtqqczze 2020-11-20
> > there is in fact a very small gain.
> 
> 0.0358 ns vs 0.6288 ns small?

Large relative gain but small absolute gain (in micro-benchmark).
> 
> It seems there are other places where we use the multiply.

This PR replaces all instances of the  `x * (y ? 1 : -1)` pattern.