@iSazonov 2020-11-20
> This PR replaces all instances of the `x * (y ? 1 : -1)` pattern.

See `private object Compare(object objValue, object statMinOrMaxValue, bool isMin)` in Measure-Object.
I guess we have others. I don't ask you to find all and fix but it would be great.