@anmenaga 2020-11-20
How the measurements were done / results table in `PR Context` was created?