@xtqqczze 2020-11-21
> How the measurements were done / results table in `PR Context` was created?

Using BenchmarkDotNet:

```
...
        bool _ascendingOrder = false;
        int result = 7;

        [Benchmark]
        public int Mul()
        {
            return result * (_ascendingOrder ? 1 : -1);
        }

        [Benchmark]
        public int Neg()
        {
            return _ascendingOrder ? result : -result;
        }
...
```

Not the most realistic benchmark, but the results do show performance is not harmed by the changes.